Validate buffer size based on packet header
Original implementation compared the computed data size (based on sizes in the packet header) to the buffer size less the nominal size of the known packet header structure.  However, some (I guess older?) R2000s use a smaller packet header (60 bytes vs 76).  Trying to parse packets from these devices results in almost universal failure, the original condition ends up being "does the packet have at least 16 extra bytes at the end" to which the answer is usually "no" (except if the packet is part of a TCP stream and has the beginning of the next PFSDP packet immediately after it in the buffer).

This code is still not ideal, as the first few data bytes are also parsed into the header in that case, so we'll end up with some header fields that might be invalid (with no indication to the user).  It also doesn't modify the outer check (in `include/pf_driver/pf/pf_parser.h`) that verifies that there's at least as much data as the expected header size, so in rare cases (packets with only a few points) we may drop (or delay) those packets; however that should be fairly uncommon (packet sizes are based on scan parameters and AFAICT tend to avoid nearly empty packets).